Spring Boot. Lesson 12. JPA oneToMany - user message, message editor.

+++ src/main/java/com/example/sweater/repos/MessageRepos.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface MessageRepos extends CrudRepository<Message, Integer> {
+public interface MessageRepos extends CrudRepository<Message, Long> {
     List<Message> findByTag(String text);
 }
